refactor: externalize data and utils
