+++ src/data/items.js
@@ -1,0 +1,9 @@
+export const ITEMS = [
+  { id: "cursor", name: "Curseur", emoji: "🖱️", base: 18, growth: 1.15, cps: 0.22, desc: "Clique automatiquement à petits pas." },
+  { id: "grandma", name: "Mamie", emoji: "👵", base: 140, growth: 1.15, cps: 1.8, desc: "Cuit des cookies maison." },
+  { id: "farm", name: "Ferme", emoji: "🌾", base: 1400, growth: 1.16, cps: 14, desc: "Culture intensive de cookies." },
+  { id: "factory", name: "Usine", emoji: "🏭", base: 16000, growth: 1.17, cps: 72, desc: "Production industrielle." },
+  { id: "bank", name: "Banque", emoji: "🏦", base: 180000, growth: 1.18, cps: 360, desc: "Intérêts en cookies." },
+  { id: "ai", name: "IA Boulangerie", emoji: "🤖", base: 2200000, growth: 1.19, cps: 1800, desc: "Réseaux de neurones pâtissiers." },
+  { id: "tm", name: "Machine à Temps", emoji: "⌛", base: 32000000, growth: 1.21, cps: 9800, desc: "Cuisson avant même d'avoir faim." },
+];
